+++ src/frontend/components/DashboardVideoLivePairing/index.tsx
@@ -1,0 +1,140 @@
+import { Meter, Text } from 'grommet';
+import React, { useEffect, useState } from 'react';
+import { defineMessages, FormattedMessage } from 'react-intl';
+
+import { DashboardButton } from 'components/DashboardPaneButtons/DashboardButtons';
+import { usePairingVideo } from 'data/queries';
+import { Video } from 'types/tracks';
+
+const messages = defineMessages({
+  pairingSecretLabel: {
+    defaultMessage: 'Pair an external device',
+    description:
+      'Button label to request a secret for pairing an external device.',
+    id: 'components.DashboardVideoLivePairing.pairingSecretLabel',
+  },
+  pairingSecretDisplay: {
+    defaultMessage: 'Pairing secret: {secret}',
+    description:
+      'Message displaying the secret for pairing an external device.',
+    id: 'components.DashboardVideoLivePairing.pairingSecretDisplay',
+  },
+  pairingSecretCountdownExpired: {
+    defaultMessage: 'Pairing secret expired',
+    description:
+      'Message displayed when the secret for pairing an external device has expired.',
+    id: 'components.DashboardVideoLivePairing.pairingSecretCountdownExpired',
+  },
+});
+
+interface DashboardVideoLivePairingProps {
+  video: Video;
+}
+
+export const DashboardVideoLivePairing = ({
+  video,
+}: DashboardVideoLivePairingProps) => {
+  // secret expiration
+  const expiration = 0;
+  // reset 3 seconds after expiration
+  const reset = -3;
+
+  const [buttonColor, setButtonColor] = useState('brand');
+  const [secret, setSecret] = useState('');
+  const [expiresIn, setExpiresIn] = useState(0);
+  const [secondsLeft, setSecondsLeft] = useState(0);
+  const [expired, setExpired] = useState(false);
+
+  const usePairingVideoMutation = usePairingVideo(video.id, {
+    onSuccess: (data) => {
+      setSecret(data.secret);
+      setExpiresIn(data.expires_in);
+      setSecondsLeft(data.expires_in);
+      setExpired(false);
+      setButtonColor('status-ok');
+    },
+  });
+
+  const pairVideoAction = () => {
+    usePairingVideoMutation.mutate();
+  };
+
+  const upddateStates = () => {
+    setSecondsLeft(secondsLeft - 1);
+
+    switch (secondsLeft) {
+      case expiration:
+        // the secret is expired, set states to display an expiration message
+        setExpired(true);
+        setButtonColor('status-warning');
+        break;
+      case reset:
+        // the secret expired 3 seconds ago, set states to display initial label
+        setSecret('');
+        setButtonColor('brand');
+    }
+  };
+
+  useEffect(() => {
+    let timeout: number | undefined;
+    // don't start a timeout if secret doesn't exists
+    // and reset has been done one second ago
+    if (secret && secondsLeft > reset - 1) {
+      timeout = window.setTimeout(upddateStates, 1000);
+    }
+    return () => {
+      window.clearTimeout(timeout);
+    };
+  }, [secondsLeft]);
+
+  let content: JSX.Element;
+
+  if (!secret) {
+    // no secret yet, initial state
+    content = (
+      <Text alignSelf="center">
+        <FormattedMessage {...messages.pairingSecretLabel} />
+      </Text>
+    );
+  } else {
+    if (!expired) {
+      // secret exists and is not expired yet
+      content = (
+        <React.Fragment>
+          <Text alignSelf="center" size="large">
+            <FormattedMessage
+              {...messages.pairingSecretDisplay}
+              values={{ secret }}
+            />
+          </Text>
+          <Meter
+            type="bar"
+            color="light-1"
+            background={{ color: 'light-1', opacity: 'medium' }}
+            value={(secondsLeft * 100) / expiresIn}
+            alignSelf="center"
+            size="full"
+            thickness="xsmall"
+            margin={{ top: 'small' }}
+          />
+        </React.Fragment>
+      );
+    } else {
+      // secret has expired
+      content = (
+        <Text alignSelf="center">
+          <FormattedMessage {...messages.pairingSecretCountdownExpired} />
+        </Text>
+      );
+    }
+  }
+
+  return (
+    <DashboardButton
+      primary
+      color={buttonColor}
+      onClick={pairVideoAction}
+      label={content}
+    />
+  );
+};